style: 중복되는 함수 util로 분리 #19
