+++ src/app/personal/[id]/page.tsx
@@ -2,21 +2,20 @@
 
 import { useEffect, useRef, useState } from 'react';
 import { axiosInstance } from '@/utils/axios';
-import { content } from '@/types/content';
 import StrcatBoard from '@/component/StrcatBoard';
-import Add from '@/component/Add';
 import BottomButton from '@/component/BottomButton';
 import ContentPhoto from '@/component/ContentPhoto';
 import { useRecoilState } from 'recoil';
-import { themeObj, themeState } from '@/recoil/theme';
+import { themeObj } from '@/recoil/theme';
 import Drawer from '@/component/Drawer';
 import StrcatHeader from '@/component/StrcatHeader';
 import { observeState } from '@/recoil/observe';
 import { useRouter } from 'next/navigation';
 import { board } from '@/types/boards';
+import { scrollToAdd, setMap } from '@/utils/scrollTo';
 
 export default function Home() {
-  const [boards, setBoards] = useState<board>({
+  const [board, setBoard] = useState<board>({
     id: 0,
     title: '',
     theme: 'strcat',
@@ -26,29 +25,18 @@ export default function Home() {
   const itemsRef = useRef(new Map());
   const [observe] = useRecoilState(observeState);
   const router = useRouter();
-
-  const scrollToId = (itemId: number) => {
-    const map = getMap();
-    const node = map.get(itemId);
-    const height = node.offsetHeight;
-    const offset = node.offsetTop + height - 500; // 하단의 여백을 500만큼 줬으므로 그만큼 빼준다.
-    window.scrollTo({ top: offset, behavior: 'smooth' });
-  };
-  const getMap = () => {
-    return itemsRef.current;
-  };
   useEffect(() => {
     axiosInstance
       .get(`/api/personal`)
       .then((data) => {
-        setBoards(data.data);
+        setBoard(data.data);
       })
       .catch((error) => {});
   }, []);
 
   const handleClick = () => {
     setIsAdd(true);
-    scrollToId(boardId);
+    scrollToAdd(board.id, itemsRef);
   };
 
   return (
@@ -57,22 +45,12 @@ export default function Home() {
       <StrcatHeader />
       <div
         className={`relative w-full  p-[24px] text-justify ${
-          themeObj[boards.theme].BgColor
+          themeObj[board.theme].BgColor
         } pb-[500px]`}
       >
         <StrcatBoard
-          theme={boards.theme}
-          ref={(node) => {
-            const map = getMap();
-            if (node) {
-              map.set(boards.id, node);
-            } else {
-              map.delete(boards.id);
-            }
-          }}
-          boardId={boards.id}
-          title={boards.title}
-          data={boards.content}
+          board={board}
+          ref={(node) => setMap(node, board, itemsRef)}
           isAdd={isAdd}
           setIsAdd={setIsAdd}
         />
